// src/types/quiz.ts
export interface Question {
  category: string;
  type: string;
  difficulty: string;
  question: string;
  correct_answer: string;
  incorrect_answers: string[];
}

export interface QuizState {
  questions: Question[];
  currentQuestionIndex: number;
  score: number;
  userAnswers: (string | null)[];
  isLoading: boolean;
  isQuizComplete: boolean;
}